Complete customer creation logic and start customer finding logic
There are a couple key requirements for customers to be registered. At most they can fill out all the fields required or the essential ones being username, email, password, if this is not present in the creation request, it must throw an exception. To do this a custom exception
was created for code clarity.