Add Allergies and Calories to MenuService
- Added: Allergies and Calories fields to menu inside MenuService
- Added: Fields to MenuItem for allergens and calories
- Updated: Menu in MenuControllerTest to new menu

MenuService has been updated to now return allergies and calories as separate fields. This supports the customer in avoiding meals that contain certain allergens and of high calorie count.

MenuItem has been updated accordingly with the new fields and getter/setter methods to support this new data, and the sample menu in MenuControllerTest has been updated to match the primary menu.