Refined Unit Tests for Menu Service
- Modified: Scope of MenuControllerTest is now MenuService class
- Renamed: MenuControllerTest.java -> TestMenuService.java
- Renamed: MenuServiceApplicationTests.java -> TestMenuServiceApplication.java
- Fixed: Unit tests passing dependent on service already running
- Added: toString() and equals() methods to MenuItem

This commit focuses on refactoring and refining the existing test cases for the menu service.

The test classes have been renamed to be consistent with the test classes in CS2800 and the scope of the MenuControllerTest class has changed to now test the MenuService class. This is because even though the tests are correct, an exception is thrown when MenuService is not running in the background. This change in scope has been reflected in the updated name of the test class.

Additionally the testGetMenuItem_AllItems() test has been fixed and now works as intended, and a toString() and equals() method have been added to MenuItem to assist in testing.